Fixed temp_file_for_content not closing file before returning path
Minor typos

+++ src/penai/utils/svg.py
@@ -8,8 +8,8 @@ from typing import Any
 from urllib.parse import urlparse
 
 import requests
-from lxml import etree
 from PIL import Image
+from lxml import etree
 
 
 def trim_namespace_from_tree(node: etree.Element, namespace: str) -> None:
@@ -54,7 +54,7 @@ def validate_uri(x: Any) -> bool:
 
 
 @contextmanager
-def temp_file_for_content(content: str | bytes, extension: str) -> Generator[Path, Any, Any]:
+def temp_file_for_content(content: str | bytes, extension: str, delete: bool = False) -> Generator[Path, Any, Any]:
     """Create a temporary file for a given file content."""
     if extension and not extension.startswith("."):
         raise ValueError("Extension should start with a dot")
@@ -65,9 +65,15 @@ def temp_file_for_content(content: str | bytes, extension: str) -> Generator[Pat
         assert isinstance(content, bytes)
         mode = "wb"
 
-    with NamedTemporaryFile(prefix="penai_", suffix=extension, mode=mode, delete=False) as file:
+    # Note: (just for the curious, not actually needed to know)
+    # buffering=0 is very important if you want to yield inside
+    # Since we don't use the delete option here (we delete manually below)
+    # we yield outside of this context
+    # The code below is essentially equivalet to wit open()..write
+    with NamedTemporaryFile(prefix="penai_", suffix=extension, mode=mode, delete=False, buffering=0) as file:
         file.write(content)
         path = Path(file.name)
-        yield path
+    yield path
 
-    path.unlink()
+    if delete:
+        path.unlink()
